requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2025.4
cryptography==44.0.2
daphne==4.2.0
distro==1.9.0
Django==5.1.7
django-allauth==65.7.0
django-cleanup==9.0.0
django-crispy-forms==2.3
django-elasticsearch-dsl==8.0
django-htmx==1.23.0
django-multiselectfield==0.1.13
django-widget-tweaks==1.5.0
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.9.0
langdetect==1.0.9
multidict==6.4.0
openai==0.28.0
pillow==11.2.1
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyArabic==0.6.15
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
zope.interface==7.2
